# Increase Llama3.1-70B and Llama2-70B benchmark batch sizes